@@ -1,24 +1,37 @@
-source 'http://rubygems.org'
+source 'https://rubygems.org'
 
-gem 'rails', '3.0.6'
-gem 'mysql2', '0.2.7'
-gem 'devise'
+gem 'rails', '3.2.1'
+
+# Bundle edge Rails instead:
+# gem 'rails', :git => 'git://github.com/rails/rails.git'
+
+gem 'mysql2'
+gem 'sqlite3'
+gem 'devise', :git => 'https://github.com/plataformatec/devise.git'
 gem 'annotate'
 gem 'formtastic'
+gem 'omniauth'
+gem 'omniauth-facebook'
 gem 'oa-oauth', :require => 'omniauth/oauth'
 gem 'paperclip', :git => 'https://github.com/thoughtbot/paperclip.git'
 gem 'haml'
 gem 'inherited_resources'
 gem 'meta_search'
-gem 'rmagick', '2.12.2'
+#gem 'rmagick', '2.12.2'
 gem 'state_machine'
-gem 'will_paginate', '~> 3.0'
+gem 'will_paginate'
 gem 'activemerchant'
 gem 'sqlite3'
 gem 'activeadmin'
 gem 'rails3-jquery-autocomplete'
-gem 'ruby-debug19', :require => 'ruby-debug'
 gem 'airbrake', '~> 3.0.9'
+gem 'jquery-rails'
+
+group :assets do
+  gem 'sass-rails',   '~> 3.2.3'
+  gem 'coffee-rails', '~> 3.2.1'
+  gem 'uglifier', '>= 1.0.3'
+end
 
 group :development, :test do
   gem 'cucumber-rails'
@@ -32,4 +45,6 @@ group :development, :test do
   gem 'launchy'
   gem 'capistrano'
   gem 'engineyard'
+  gem 'ruby-debug19', :require => 'ruby-debug'
 end
+
