@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Sprowty::Application.config.secret_token = '7c5e9321a427e890141759d61c6ccae056e12db1db41cf3d3d118b8283f87348d5acb374fb6b684c68181eccb9b22efad68af81de7ecd8e166f50b1da8a9ea16'
+Sprowty::Application.config.secret_token = 'c862eee0e3068930a4a5fc785b1cc7149514e69ac98a1fb689e60dd3942bbddedd928f1b267a9ced6e457bb90be05c94bbb64f2db6858c4548e7b3e25ac0f792'
