@@ -1,3 +1,4 @@
+# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -29,8 +30,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.integer  "author_id"
     t.string   "author_type"
     t.text     "body"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",    :null => false
+    t.datetime "updated_at",    :null => false
     t.string   "namespace"
   end
 
@@ -48,8 +49,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                          :null => false
+    t.datetime "updated_at",                                          :null => false
   end
 
   add_index "admin_users", ["email"], :name => "index_admin_users_on_email", :unique => true
@@ -61,8 +62,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.float    "price"
     t.text     "why"
     t.datetime "offer_expires"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                              :null => false
+    t.datetime "updated_at",                              :null => false
     t.boolean  "accept_current_price"
     t.boolean  "accepted",             :default => false
   end
@@ -79,22 +80,22 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.text     "tags"
     t.float    "price"
     t.datetime "due_date"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",  :null => false
+    t.datetime "updated_at",  :null => false
   end
 
   create_table "feedbacks", :force => true do |t|
     t.integer  "project_id"
     t.integer  "user_id"
     t.text     "feedback"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
   end
 
   create_table "keywords", :force => true do |t|
     t.string   "name"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
   end
 
   create_table "message_recipients", :force => true do |t|
@@ -117,8 +118,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.string   "state",       :null => false
     t.datetime "hidden_at"
     t.string   "type"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",  :null => false
+    t.datetime "updated_at",  :null => false
   end
 
   create_table "problems", :force => true do |t|
@@ -126,8 +127,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.integer  "project_id"
     t.string   "name"
     t.string   "contact"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                     :null => false
+    t.datetime "updated_at",                     :null => false
     t.string   "status",     :default => "open"
   end
 
@@ -139,8 +140,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.string   "paypal"
     t.string   "facebook"
     t.string   "twitter"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                              :null => false
+    t.datetime "updated_at",                              :null => false
     t.integer  "user_id"
     t.string   "picture_file_name"
     t.string   "picture_content_type"
@@ -169,8 +170,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.text     "tags"
     t.float    "price"
     t.datetime "due_date"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                     :null => false
+    t.datetime "updated_at",                     :null => false
     t.integer  "bid_count",   :default => 0
     t.string   "city"
     t.string   "state"
@@ -191,22 +192,22 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.text     "education"
     t.text     "sample_work"
     t.integer  "user_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",  :null => false
+    t.datetime "updated_at",  :null => false
   end
 
   create_table "skills", :force => true do |t|
     t.integer  "user_id"
     t.string   "skill"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
   end
 
   create_table "transactions", :force => true do |t|
     t.integer  "account_id"
     t.float    "amount"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",            :null => false
+    t.datetime "updated_at",            :null => false
     t.integer  "paypal_transaction_id", :null => false
     t.string   "sm_state"
     t.string   "tx_type"
@@ -227,8 +228,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.integer  "failed_attempts",                     :default => 0
     t.string   "unlock_token"
     t.datetime "locked_at"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                             :null => false
+    t.datetime "updated_at",                                             :null => false
     t.string   "username"
     t.boolean  "is_admin",                            :default => false
     t.string   "confirmation_token"
@@ -248,8 +249,8 @@ ActiveRecord::Schema.define(:version => 20120115035359) do
     t.string   "work_content_type"
     t.integer  "work_file_size"
     t.datetime "work_updated_at"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                           :null => false
+    t.datetime "updated_at",                           :null => false
     t.boolean  "image_approved",    :default => false
   end
 
